pyproject.toml
[tool.poetry]
name = "wagtail-graphql"
version = "0.2.0"
description = "An app to automatically add graphql support to a Wagtail website"
license = "MIT"
readme = "README.md"
authors = [
    "Tiago Requeijo <tiago.requeijo.dev@gmail.com>"
]
repository = "https://github.com/tr11/wagtail-graphql"
homepage = "https://github.com/tr11/wagtail-graphql"
keywords = ['wagtail', 'graphql', 'api', 'wagtail-graphql']

[tool.poetry.dependencies]
python = "~3.6||~3.7"
graphene_django = "^2.2"
graphene_django_optimizer = "^0.3.5"
wagtail = "^2.3"
wagtailmenus = {version = "^2.12", optional = true}
python-dateutil = "^2.7"

[tool.poetry.extras]
menus = ["wagtailmenus"]

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-django = "^3.4"
pytest-cov = "^2.6"
flake8 = "^3.6"
pytest-pythonpath = "^0.7.3"
mypy = "^0.641.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
